Agregado de cálculo de ganancias y gestión de fechas
- Se agregó el cálculo de Ganancia Total sin filtrar por fecha.
- Se implementaron filtros de fecha para ganancias anual, mensual y diaria.
- Se aseguró que las órdenes futuras no afecten el cálculo diario.
- Se añadió la funcionalidad para actualizar las ganancias automáticamente al cambiar de mes.
- Se mejoró la estructura del código para facilitar futuras modificaciones.

+++ backend/src/main/java/com/gestor/tienda/Dto/DetalleOrdenDto.java
@@ -18,7 +18,7 @@ public class DetalleOrdenDto {
     private int ordenId;
 
     @NotNull
-    private int productoId;
+    private long productoId;
 
     @NotNull
     private int cantidad;
